Use `didPrepare` for zipping FastBoot-build
This makes sure that we zip the build later than the creation
of the revision-key which is usually done on `prepare` via the
revision-data-plugin.

This makes it easier to configure the plugin for normal users
as they don't have to specify plugin run-order for this plugin
to do the right thing.